requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
aiosqlite==0.19.0
aiohttp==3.9.1
yookassa==2.4.0
